Gracefully handle files with no statements
This use case came up recently in a set of tests that generated a dummy module. The dummy module was getting picked up by Devel::Cover but then it had no statement level information so `$f->statement` is just `undef` and the call to `$st->items` dies. I'm separately solving this problem by filtering those out of the cover report to begin with but this seemed like a reasonable fix to handle it more gracefully.